refactor(kitchenpos): separate domains from classes into interfaces
