+++ src/main/java/kitchenpos/domain/MenuRepository.java
@@ -1,13 +1,17 @@
 package kitchenpos.domain;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
-public interface MenuRepository extends JpaRepository<Menu, UUID> {
-    @Query("select m from Menu m, MenuProduct mp where mp.product.id = :productId")
-    List<Menu> findAllByProductId(@Param("productId") UUID productId);
+public interface MenuRepository {
+    Menu save(Menu menu);
+
+    Optional<Menu> findById(UUID uuid);
+
+    List<Menu> findAll();
+
+    List<Menu> findAllById(List<UUID> uuids);
+
+    List<Menu> findAllByProductId(UUID productId);
 }
