+++ src/main/java/kitchenpos/domain/ProductRepository.java
@@ -1,8 +1,15 @@
 package kitchenpos.domain;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
+import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
-public interface ProductRepository extends JpaRepository<Product, UUID> {
+public interface ProductRepository {
+    Product save(Product product);
+
+    Optional<Product> findById(UUID uuid);
+
+    List<Product> findAll();
+
+    List<Product> findAllById(List<UUID> uuids);
 }
